Refactored presentation.
