+++ src/business/gateways/Movies/newMovieGateway.ts
@@ -1,0 +1,6 @@
+import { Movie } from "../../entities/Movie";
+
+
+export interface newMovieGateway {
+    saveMovie(movie: Movie): Promise<void>
+};
