+++ src/business/gateways/Movies/searchMovieGateway.ts
@@ -1,0 +1,3 @@
+export interface searchMovieGateway {
+    getMovie(): Promise<void>
+};
