--- src/business/gateways/newMovieGateway.ts
@@ -1,7 +1,0 @@
-import { Movie } from "../entities/Movie";
-import knex from 'knex'
-
-
-export interface newMovieGateway {
-    saveMovie(movie: Movie): Promise<knex>
-};
