+++ src/business/usecases/Movies/searchMovie/searchMovie.ts
@@ -1,0 +1,10 @@
+import { searchMovieGateway } from './../../../gateways/Movies/searchMovieGateway';
+
+
+export class SearchMovieUseCase {
+    constructor(private gateway: searchMovieGateway ){}
+
+    async execute() {
+        return await this.gateway.getMovie()
+    }
+}
